% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBO2019-package.R
\docType{package}
\name{KBO2019-package}
\alias{KBO2019}
\alias{KBO2019-package}
\title{KBO2019: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MrKevinNa/KBO2019}
  \item Report bugs at \url{https://github.com/MrKevinNa/KBO2019/issues}
}

}
\keyword{internal}
